Refactor user module structure
