+++ .eslintrc.json
@@ -13,7 +13,8 @@
     "no-useless-constructor": "off",
     "no-empty-function": "off",
     "import/prefer-default-export": "off",
-    "import/extensions": "off"
+    "import/extensions": "off",
+    "linebreak-style": "off"
   },
   "settings": {
     "alias": {
