+++ src/core/interfaces/IService.ts
